requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
django-filter
pillow
djangorestframework-simplejwt
requests 
beautifulsoup4
lxml